reorganize token components into a folder
