+++ apps/frontend/components/token/stats.tsx
